UIIN-1868: Moved MarcView component to quickMARC module
