+++ src/Holding/HoldingsMarc/HoldingsMarcContainer.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 
 import { stripesConnect } from '@folio/stripes/core';
 
-import { MarcContainer } from '../../components';
+import { ViewSource } from '../../components';
 
 const HoldingsMarcContainer = ({
   mutator,
@@ -10,7 +10,7 @@ const HoldingsMarcContainer = ({
   holdingsrecordid,
 }) => {
   return (
-    <MarcContainer
+    <ViewSource
       mutator={mutator}
       instanceId={instanceId}
       holdingsRecordId={holdingsrecordid}
