--- src/components/MarcView/MarcView.css
@@ -1,14 +1,0 @@
-.marcWrapper {
-  margin-left: 20px;
-}
-
-.marc {
-  table-layout: fixed;
-  border: 0;
-  font-family: courier;
-  white-space: pre-wrap;
-
-  & tr td {
-    vertical-align: top;
-  }
-}
