--- src/components/MarcView/index.js
@@ -1,1 +1,0 @@
-export { default } from './MarcContainer';
