--- src/components/MarcView/utils.js
@@ -1,3 +1,0 @@
-export const isControlField = field => (Object.keys(field)[0]).startsWith('00');
-
-export const normalizeIndicator = indictor => indictor.replace(/\\/g, ' ');
