+++ src/components/index.js
@@ -14,4 +14,4 @@ export { default as ModalContent } from './ModalContent';
 export { PaneLoading, ViewLoading } from './Loading';
 export { default as WarningMessage } from './WarningMessage';
 export { default as CalloutRenderer } from './CalloutRenderer';
-export { default as MarcContainer } from './MarcView';
+export { default as ViewSource } from './ViewSource';
